Remove duplicate code and files
Remove duplicate code and consolidate functionality across the project.

* **PrismaClient Initialization**: Add `src/lib/prismaClient.ts` to initialize and export `PrismaClient`. Modify `src/features/analytics/services/behaviorAnalyticsService.ts`, `src/features/analytics/services/reportingService.ts`, and `src/features/analytics/services/usageStatsService.ts` to import `prisma` from the new shared file and remove local `PrismaClient` initialization.
* **User Registration**: Delete `src/pages/api/auth/signup.ts` and update all references to use `src/pages/api/auth/register.ts`.
* **useAuth Hook**: Remove duplicate `useAuth` hook in `src/hooks/useAuth.tsx` and update all references to use `src/features/auth/hooks/useAuth.tsx`. Modify `src/features/auth/hooks/useAuth.tsx` to add `token` to `AuthContextType` interface and `AuthContext.Provider` value.
* **Login Functionality**: Modify `src/pages/api/auth/login.ts` to import `login` from `src/features/auth/services/authService.ts` and replace existing login logic with `login` function call.
* **User Data Fetching**: Modify `src/pages/api/auth/user.ts` to import `getCurrentUser` from `src/features/auth/services/authService.ts` and replace existing user fetching logic with `getCurrentUser` function call.
