--- src/pages/api/auth/signup.ts
@@ -1,35 +1,0 @@
-import { NextApiRequest, NextApiResponse } from 'next'
-import { PrismaClient } from '@prisma/client'
-import bcrypt from 'bcryptjs'
-
-const prisma = new PrismaClient()
-
-export default async function handler(req: NextApiRequest, res: NextApiResponse) {
-  if (req.method !== 'POST') {
-    return res.status(405).json({ message: 'Method not allowed' })
-  }
-
-  try {
-    const { name, email, password } = req.body
-
-    const existingUser = await prisma.user.findUnique({ where: { email } })
-    if (existingUser) {
-      return res.status(400).json({ message: 'User already exists' })
-    }
-
-    const hashedPassword = await bcrypt.hash(password, 10)
-    const user = await prisma.user.create({
-      data: {
-        name,
-        email,
-        password: hashedPassword,
-      },
-    })
-
-    res.status(201).json({ message: 'User created successfully', userId: user.id })
-  } catch (error) {
-    console.error('Signup error:', error)
-    res.status(500).json({ message: 'Error creating user' })
-  }
-}
-
